feat: Add formatter for console interface
Improve user experience with a clean and interactive console interface.
The `ConsoleFormatter` class uses the `Rich` library to show weather data
and provide a user-friendly menu. This makes the console output look
nice and easy to use, which helps to improve the overall usability
and functionality of the weather app.