refactor: update weather service to use Pydantic
* Use Pydantic models to check and parse data more easily.
* Add OpenWeatherSchema to match OpenWeather API response.
* Change service methods to use WeatherDTO with Pydantic.
* Make weather type logic simpler and easier to read.
* Improve _parse_* methods for better error handling.
* Update API response handling to match the new model.

These changes were made to make the weather service and models
easier to work with. We replaced the old model with Pydantic-based
models. This helps to check and process the data better.

The new OpenWeatherSchema is used to organize the data from the
OpenWeather API. This makes the code easier to manage.